New class StrainTriples
